pyproject.toml
[project]
name = "pdf2zh"
version = "1.9.10"
description = "Latex PDF Translator"
authors = [{ name = "Byaidu", email = "byaidux@gmail.com" }]
license = "AGPL-3.0"
readme = "README.md"
requires-python = ">=3.10,<3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    # for arm64 linux whells
    "pymupdf<1.25.3",
    "tqdm",
    "tenacity",
    "numpy",
    "ollama",
    "xinference-client",
    "deepl",
    "openai>=1.0.0",
    "azure-ai-translation-text<=1.0.1",
    "gradio",
    "huggingface_hub",
    "onnx",
    "onnxruntime",
    "opencv-python-headless",
    "pytesseract",
    "pillow",
    "tencentcloud-sdk-python-tmt",
    "pdfminer-six==20250416",
    "gradio_pdf>=0.0.21",
    "pikepdf",
    "peewee>=3.17.8",
    "fontTools",
    "babeldoc>=0.1.22, <0.3.0",
    "rich",
]

[project.optional-dependencies]
backend = [
    "flask",
    "celery",
    "redis"
]
argostranslate = [
    "argostranslate"
]
mcp = [
    "mcp>=1.6.0",
]

[dependency-groups]
dev = [
    "black",
    "flake8",
    "pre-commit",
    "pytest",
    "build",
    "bumpver>=2024.1130",
]

[project.urls]
Homepage = "https://github.com/Byaidu/PDFMathTranslate"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
pdf2zh = "pdf2zh.pdf2zh:main"

[tool.flake8]
ignore = ["E203", "E261", "E501", "W503", "E741"]
max-line-length = 88



[bumpver]
current_version = "1.9.10"
version_pattern = "MAJOR.MINOR.PATCH[.PYTAGNUM]"

[bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]
"pdf2zh/__init__.py" = [
    '__version__ = "{version}"'
]
